// frontend/src/lib/api/generated/models/ItemAvailabilityResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemAvailabilityResponse = {
    itemId: string;
    totalQuantity: number | string;
    reservedOrRentedQuantity: number | string;
    availableQuantity: number | string;
    from: string;
    to: string;
};